refactor: divide code in several files
