+++ apps/front/src/components/coaches/table/index.tsx
@@ -1,69 +1,58 @@
-"use client";
-import React, { useState } from 'react';
-import {
-  Table,
-  TableCaption,
-} from "@/components/ui/table"
-import { MagnifyingGlassIcon } from '@radix-ui/react-icons';
-import { Coach } from '..';
+import { Table } from "@/components/ui/table";
+import { Coach } from "..";
+import { ColumnFiltersState, getCoreRowModel, getFilteredRowModel, getPaginationRowModel, getSortedRowModel, SortingState, useReactTable, VisibilityState } from "@tanstack/react-table";
+import { useEffect, useState } from "react";
+import CoachTableFooter from "./footer";
+import { columns } from "./config";
+import CoachesTableBody from "./body";
+import CoachesTableHeader from "./header";
+import CoachesTableFilters from "./filter";
 
-export interface CoachesListProps {
+export interface CoachesTableProps {
   coaches: Coach[];
 }
 
-export type Action = 'delete' | 'promote' | 'demote';
+export default function CoachesTable(
+  { coaches }: CoachesTableProps
+) {
+  const [sorting, setSorting] = useState<SortingState>([])
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([])
+  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({})
+  const [rowSelection, setRowSelection] = useState({})
 
-export default function CoachesTable({ coaches }: CoachesListProps) {
-  const [displayedCoaches, setDisplayedCoaches] = useState<Coach[]>(coaches);
-  const [selectedCoaches, setSelectedCoaches] = useState<Coach[]>([]);
+  const table = useReactTable({
+    data: coaches,
+    columns,
+    onSortingChange: setSorting,
+    onColumnFiltersChange: setColumnFilters,
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    onColumnVisibilityChange: setColumnVisibility,
+    onRowSelectionChange: setRowSelection,
+    state: {
+      sorting,
+      columnFilters,
+      columnVisibility,
+      rowSelection,
+    },
+  })
 
-  const handleSelectAllCoaches = () => {
-    if (selectedCoaches.length === coaches.length) {
-      setSelectedCoaches([]);
-    } else {
-      setSelectedCoaches(coaches);
-    }
-  }
-
-  const handleSelectCoach = (coach: Coach) => {
-    if (selectedCoaches.includes(coach)) {
-      setSelectedCoaches(selectedCoaches.filter((selectedCoach) => selectedCoach !== coach));
-    } else {
-      setSelectedCoaches([...selectedCoaches, coach]);
-    }
-  }
-
-  const handleSearch = (search: string) => {
-    if (search === '') {
-      setDisplayedCoaches(coaches);
-    } else {
-      setDisplayedCoaches(coaches.filter((coach) => (
-        coach.firstName.toLowerCase().includes(search.toLowerCase()) ||
-        coach.lastName.toLowerCase().includes(search.toLowerCase()) ||
-        coach.email.toLowerCase().includes(search.toLowerCase()) ||
-        coach.phone.toLowerCase().includes(search.toLowerCase())
-      )));
-    }
-  }
-
-  const handleAction = (action: Action) => {
-    if (action === 'delete') {
-      console.log('Delete', selectedCoaches);
-    } else if (action === 'promote') {
-      console.log('Promote', selectedCoaches);
-    } else if (action === 'demote') {
-      console.log('Demote', selectedCoaches);
-    }
-  }
+  useEffect(() => {
+    console.log(rowSelection)
+  }, [rowSelection])
 
   return (
-    <div>
-      <div>
-        <MagnifyingGlassIcon className="w-6 h-6" />
+    <div className="w-full">
+      <CoachesTableFilters table={table} />
+      <div className="rounded-md border">
+        <Table>
+          <CoachesTableHeader table={table} />
+          <CoachesTableBody table={table} />
+        </Table>
       </div>
-      <Table>
-        <TableCaption>{coaches.length} coaches are displayed</TableCaption>
-      </Table>
+      <CoachTableFooter table={table} />
     </div>
-  );
+  )
 }
